Add a prefs widget with an about page
We use a separate typescript config to isolate the prefs file from the
extension files; tsc compiles all files reachable from a single tsconfig
into a shared global namespaces whereas the prefs widget is normally
entirely indepdent.

The prefs widget doesn't do much since we have no settings, but it
serves as a place to show an About page, with license information and a
link to our Github page.

Also split the type decl files, move them to a subdirectory, and move
them to a subdirectory, for easier syncing with vscode search provider.